port to 1.21.11

+++ common/src/main/java/de/keksuccino/panoramica/mixin/mixins/common/client/MixinGameRenderer.java
@@ -12,9 +12,9 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public class MixinGameRenderer {
 
     @Inject(method = "getFov", at = @At("HEAD"), cancellable = true)
-    private void headGetFov_Panoramica(Camera camera, float f, boolean bl, CallbackInfoReturnable<Double> info) {
+    private void headGetFov_Panoramica(Camera camera, float f, boolean bl, CallbackInfoReturnable<Float> info) {
         if (PanoramicaHandler.panoramaMode) {
-            info.setReturnValue(90.0D);
+            info.setReturnValue(90.0F);
         }
     }
